feat(controllers): add return types and validation to methods

+++ code/server/src/controllers/reviewController.ts
@@ -1,22 +1,32 @@
 import { User } from "../components/user";
 import ReviewDAO from "../dao/reviewDAO";
+import {ProductReview} from "../components/review";
+import {ExistingReviewError, NoReviewProductError} from "../errors/reviewError";
+import ProductDAO from "../dao/productDAO";
+import {ProductNotFoundError} from "../errors/productError";
 
 class ReviewController {
     private dao: ReviewDAO
+    private productDao: ProductDAO
 
     constructor() {
         this.dao = new ReviewDAO
+        this.productDao = new ProductDAO
     }
 
     /**
      * Adds a new review for a product
      * @param model The model of the product to review
-     * @param user The username of the user who made the review
+     * @param user The user who made the review
      * @param score The score assigned to the product, in the range [1, 5]
      * @param comment The comment made by the user
      * @returns A Promise that resolves to nothing
      */
-    async addReview(model: string, user: User, score: number, comment: string) /**:Promise<void> */ {
+    async addReview(model: string, user: User, score: number, comment: string): Promise<void> {
+        const product = await this.productDao.getProductByModel(model)
+        if(!product) throw new ProductNotFoundError()
+        const hasReviewed = await this.dao.hasReviewed(model, user.username)
+        if(hasReviewed) throw new ExistingReviewError()
         return this.dao.addReview(model, user, score, comment)
     }
 
@@ -25,7 +35,9 @@ class ReviewController {
      * @param model The model of the product to get reviews from
      * @returns A Promise that resolves to an array of ProductReview objects
      */
-    async getProductReviews(model: string) /**:Promise<ProductReview[]> */ {
+    async getProductReviews(model: string) :Promise<ProductReview> {
+        const product = await this.productDao.getProductByModel(model)
+        if(!product) throw new ProductNotFoundError()
         return this.dao.getProductReviews(model)
     }
 
@@ -35,7 +47,11 @@ class ReviewController {
      * @param user The user who made the review to delete
      * @returns A Promise that resolves to nothing
      */
-    async deleteReview(model: string, user: User) /**:Promise<void> */ {
+    async deleteReview(model: string, user: User): Promise<void>{
+        const product = await this.productDao.getProductByModel(model)
+        if(!product) throw new ProductNotFoundError()
+        const hasReviewed = await this.dao.hasReviewed(model, user.username)
+        if(!hasReviewed) throw new NoReviewProductError()
         return this.dao.deleteReview(model, user)
     }
 
@@ -44,7 +60,9 @@ class ReviewController {
      * @param model The model of the product to delete the reviews from
      * @returns A Promise that resolves to nothing
      */
-    async deleteReviewsOfProduct(model: string) /**:Promise<void> */ {
+    async deleteReviewsOfProduct(model: string): Promise<void>  {
+        const product = await this.productDao.getProductByModel(model)
+        if(!product) throw new ProductNotFoundError()
         return this.dao.deleteReviewsOfProduct(model)
     }
 
@@ -52,7 +70,7 @@ class ReviewController {
      * Deletes all reviews of all products
      * @returns A Promise that resolves to nothing
      */
-    async deleteAllReviews() /**:Promise<void> */ {
+    async deleteAllReviews(): Promise<void>  {
         return this.dao.deleteAllReviews()
     }
 }
